requirements.txt
fastapi==0.128.0
uvicorn==0.35.0
gunicorn==22.0.0
pandas==2.3.3
numpy==2.4.1
biopython==1.86
